[Tizen] Temporarily disable WebRTC support.
The WebRTC code underwent several changes in M37, and some of them make
its build work only if X and OpenSSL's headers are present. We are still
investigating if there is really a reason for this, but meanwhile we
need to disable WebRTC support to allow the M37 rebase to proceed.

BUG=XWALK-2160